Feat: 피격 시스템 구현
플레이어, 적 총알 모두 Bullet 클래스 상속 받음. TargetTag를 통해 구분.

+++ Assets/Scripts/IDamageable.cs
@@ -4,9 +4,9 @@ using UnityEngine;
 
 public interface IDamageable
 {
-    int Hp { get; set; }
+    float Hp { get; set; }
 
-    void Damage(int damageAmount);
+    void Damage(float damageAmount);
 
     void Die();
 }
